Major changes to the bot

+++ neo/ext/customisation.py
@@ -36,11 +36,10 @@ from yarl import URL
 
 
 class Reminder:
-    def __init__(self, *, user, bot, content, deadline, conn_pool, rm_id, jump_origin):
+    def __init__(self, *, user, bot, content, deadline, rm_id, jump_origin):
         self.user = user
         self.content = content
         self.deadline = deadline
-        self.conn_pool = conn_pool
         self.rm_id = rm_id
         self.jump_origin = jump_origin
         self.bot = bot
@@ -55,26 +54,17 @@ class Reminder:
         await self._do_remind()
 
     async def _do_remind(self):
-        target = (
-            self.bot.get_channel(int(list(URL(self.jump_origin).parts)[3])) or self.user
-        )
+        target = self.bot.get_channel(int(list(URL(self.jump_origin).parts)[3])) or self.user
         if self.user is None:
-            await self.conn_pool.execute(
-                "DELETE FROM reminders WHERE id=$1",
-                self.rm_id,
-            )
-            return
-        if self.bot.user_cache[self.user.id]["dm_reminders"] is True:
+            return await self.bot.pool.execute("DELETE FROM reminders WHERE id=$1", self.rm_id)
+        
+        if self.bot.user_cache[self.user.id].get("dm_reminders", False) is True:
             target = self.user
         send_content = f"**{self.user.mention} - <{self.jump_origin}>**\n" + indent(
             self.content, "> "
         )
-        await target.send(
-            send_content, allowed_mentions=discord.AllowedMentions(users=[self.user])
-        )
-        await self.conn_pool.execute(
-            "DELETE FROM reminders WHERE id=$1 AND user_id=$2", self.rm_id, self.user.id
-        )
+        await target.send(send_content, allowed_mentions=discord.AllowedMentions(users=[self.user]))
+        await self.bot.pool.execute("DELETE FROM reminders WHERE id=$1 AND user_id=$2", self.rm_id, self.user.id)
 
 
 class Customisation(commands.Cog):
@@ -108,7 +98,7 @@ class Customisation(commands.Cog):
                 )
             await self.bot.user_cache.refresh()
             return
-        embed = neo.Embed(title=f"""{ctx.author}'s Settings""")
+        embed = discord.Embed(title=f"""{ctx.author}'s Settings""")
         readable_settings = []
         for k, v in self.bot.user_cache[ctx.author.id].items():
             if k.startswith("_"):
@@ -149,7 +139,7 @@ class Customisation(commands.Cog):
                 )
             await self.bot.guild_cache.refresh()
             return
-        embed = neo.Embed(title=f"""{ctx.guild}'s Settings""")
+        embed = discord.Embed(title=f"""{ctx.guild}'s Settings""")
         readable_settings = []
         for k, v in self.bot.guild_cache[ctx.guild.id].items():
             if isinstance(v, bool):
@@ -167,7 +157,7 @@ class Customisation(commands.Cog):
     async def _prefix(self, ctx):
         """Invoked by itself, it will show base prefixes, and custom prefixes for the current guild
         Its subcommands delve into customisation of said prefixes"""
-        embed = neo.Embed()
+        embed = discord.Embed()
         guild_data = None
         always_active = [ctx.me.mention]
         if ctx.guild:
@@ -224,7 +214,7 @@ class Customisation(commands.Cog):
             )
         )
         await ctx.send(
-            embed=neo.Embed(description="\n".join(map(format_hl, enumerate(my_hl, 1))))
+            embed=discord.Embed(description="\n".join(map(format_hl, enumerate(my_hl, 1))))
             .set_footer(text=f"{len(my_hl)}/10 slots used")
             .set_author(
                 name=f"{ctx.author}'s highlights",
@@ -255,7 +245,7 @@ class Customisation(commands.Cog):
         await ctx.paginate(
             todos,
             10,
-            template=neo.Embed().set_author(
+            template=discord.Embed().set_author(
                 name=f"{ctx.author}'s todos ({len(todos):,} items)",
                 icon_url=ctx.author.avatar_url_as(static_format="png"),
             ),
@@ -313,7 +303,7 @@ class Customisation(commands.Cog):
 
         SELECT * FROM enumerated WHERE enumerated.rnum=$2"""
         todo = await self.bot.pool.fetchrow(query, ctx.author.id, todo_index)
-        embed = neo.Embed(description=todo["content"])
+        embed = discord.Embed(description=todo["content"])
         embed.set_footer(
             text=f"Created on {todo['created_at']:%a, %b, %d, %Y at %X UTC}"
         )
@@ -386,7 +376,7 @@ class Customisation(commands.Cog):
         await ctx.paginate(
             reminders or ["No reminders"],
             5,
-            template=neo.Embed().set_author(
+            template=discord.Embed().set_author(
                 name=ctx.author, icon_url=ctx.author.avatar_url_as(static_format="png")
             ),
             delete_on_button=True,
